Bug Description:  Modlist generates a replace operation such that:
[(1, 'nsslapd-cachememsize', None), (0, 'nsslapd-cachememsize', ['1'])]

Because this is a system configuration attribute 389 DS blocks the delete
operation, which causes the replace to fail.

Fix Description:  modlist should take a configurable argument to determine
if the resultant modlist takes the delete -> add construction, or if it should
use a ldap.MOD_REPLACE.

This value is determined by replace_is_add_then_delete. It defaults to True to
maintain the existing behaviour.