Checking for base64 encoding in ldap fails.
There are two issues here.

First, if we are checking the dn, pyldap returns this as utf8, not bytes. So we
should *always* exclude dn, because we never want to b64 it anyway.

Second, there is an issue with the safe_regex_re. It is being replaced with a
string match by some other part of the library (I cannot find where).

As a result, this causes the match to fail. We should re-compile each time with
.match on the correct bytes pattern.